Proyecto_2/cluster/redis/requirements.txt
flask
simplejson
redis
flask-cors
mysql-connector-python
python-dotenv